refactor(tests): extract next/dynamic mock to shared helper; restore IntersectionObserver after each test
- Add src/shared/utils/mocks/dynamic.tsx with reusable dynamicMock
  (parallels the existing motion mock pattern)
- Update navbar.test.tsx and contact.test.tsx to import from shared mock
  instead of duplicating the implementation
- Add afterEach(vi.unstubAllGlobals) in contact.test.tsx to prevent
  the IntersectionObserver stub from leaking into other test files
- Fix no-multi-comp lint error in navbar.tsx: extract MobileDrawerLoading
  to its own file (mobile-drawer-loading.tsx)
