+++ web/src/features/home/components/contact.test.tsx
@@ -1,32 +1,13 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import type { ComponentType } from "react";
-import { useEffect, useState } from "react";
-import { beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { testWrapper as wrapper } from "#shared/utils/tests";
 
 import { Contact } from "./contact";
 
 vi.mock(import("next/dynamic"), async () => {
-  const dynamicMock = {
-    default: (factory: () => Promise<ComponentType<Record<string, unknown>>>) => {
-      const Wrapper = (props: Record<string, unknown>) => {
-        const [Comp, setComp] = useState<ComponentType<Record<string, unknown>> | undefined>(
-          undefined,
-        );
-        useEffect(() => {
-          factory()
-            .then((mod) => {
-              setComp(() => mod);
-            })
-            .catch(() => {}); // eslint-disable-line @typescript-eslint/no-empty-function
-        }, []);
-        return Comp ? <Comp {...props} /> : undefined;
-      };
-      return Wrapper;
-    },
-  };
+  const { dynamicMock } = await import("#shared/utils/mocks/dynamic");
   return dynamicMock as unknown as Awaited<typeof import("next/dynamic")>;
 });
 
@@ -51,6 +32,10 @@ beforeEach(() => {
       public unobserve = vi.fn<() => void>();
     },
   );
+});
+
+afterEach(() => {
+  vi.unstubAllGlobals();
 });
 
 describe("Contact component", () => {
